+++ R/utils.R
@@ -115,8 +115,10 @@ open_green   <- function() if (has_crayon()) open_style("green")
 open_yellow  <- function() if (has_crayon()) open_style("yellow")
 open_magenta <- function() if (has_crayon()) open_style("magenta")
 open_cyan    <- function() if (has_crayon()) open_style("cyan")
+open_bold    <- function() if (has_crayon()) open_style("bold")
 close_colour <- function() if (has_crayon()) "\u001b[39m"
 close_italic <- function() if (has_crayon()) "\u001b[23m"
+close_bold   <- function() if (has_crayon()) close_style("bold")
 
 open_yellow_italic   <- function() if (has_crayon()) "\u001b[33m\u001b[3m"
 open_blurred_italic  <- function() if (has_crayon()) "\u001b[2m\u001b[3m"
@@ -390,4 +392,13 @@ is_testing <- function() {
 
 glue_escape <- function(x) {
   gsub("\\}", "}}", gsub("\\{", "{{", x))
+}
+
+detect_run_starts <- function(x) {
+  if (!length(x)) {
+    return(lgl())
+  }
+  starts <- x != c(NA, x[-length(x)])
+  starts[[1]] <- TRUE
+  starts
 }